CHROMIUM: cpufreq: cpu-boost: Add cpu-boost driver
When certain bursty and important events take place, it might take a
while for the current cpufreq governor to notice the new load and react
to it.  That would result in poor user experience. To alleviate this,
the cpu-boost driver tracks user interaction with the system (by
monitoring input events from keyboard, trackpad, touchscreen, etc.) and
boosts the frequency of CPUs for a short duration to maintain good user
experience while the governor catches up.

The driver is loosely based on cpu-boost.c driver from Android tree.

BUG=chrome-os-partner:58667
TEST=enable cpuboost target frequency percentage via module parameter,
monitor frequency switches on input activity

Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: Ib0718cfcc59190c7eb020c7c807d62cc53820857
Reviewed-on: https://chromium-review.googlesource.com/422418
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Conflicts:
	drivers/cpufreq/Makefile
[rebase412(groeck): Context conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
[rebase54(groeck):
	The CPUFREQ_ADJUST notifier callback is no longer supported.
	The related notifier code was removed.
	The respective code may have to be reworked. See upstream
	commit 5130802ddbb10a73cf17d1117e2044c8d5ba7d65 for details.
	]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
[rebase510(groeck): Squashed:
    CHROMIUM: cpufreq: cpu-boost: Adapt the driver to the new world
    CHROMIUM: cpufreq: cpu-boost: Fix MODULE_PARM_DESC typo
]
Signed-off-by: Guenter Roeck <groeck@chromium.org>