vb2: introduce buf_prepare/finish_for_cpu
This splits the buf_prepare and buf_finish actions into two: one
called while the cpu can still access the buffer contents, and one where
the memory has been prepared for DMA and the cpu no longer can access it.

With this change the old buf_finish is really buf_finish_for_cpu and so the
few drivers that use it are updated.

The reason for this split is that some drivers need to modify the buffer,
either before or after the DMA has taken place, in order to e.g. add JPEG
headers or do other touch ups.

You cannot do that in buf_prepare since at that time the buffer is already
synced for DMA and the CPU shouldn't touch is. So add these extra ops to
make this explicit.

Note that the dma-sg memory model doesn't sync the buffers yet in the memop
prepare. This will change in future patches.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>