CHROMIUM: block: partitions: efi: Add support for IGNOREME GPT signature
This patch adds support for a special GPT header signature marker (using
the string 'IGNOREME' instead of the spec's 'EFI PART'). This tells the
kernel to ignore this GPT completely and look at the other one instead.
Since the kernel always prefers the primary GPT anyway, all we really
need to do effectively is to check whether the primary GPT is marked
'IGNOREME' and force evaluation of the secondary one in that case.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/340080
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>

An earlier attempt to upstream an alternative solution:
https://patchwork.kernel.org/patch/8933841/

mka@: added missing param 'lba' to log message and resolved minor
conflicts.

BUG=chromium:941638, chrome-os-partner:52595
TEST=Booted on Minnie with 'IGNOREME' primary GPT

Change-Id: Ibaee639fac9fa2f04b8836ef153c95b7d0b045a4
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1531701
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>

[rebase515(groeck):
  Squash:
    FIXUP: partitions/efi: Support non-standard GPT location
    (API change)
]
Signed-off-by: Guenter Roeck <groeck@google.com>