drivers: mailbox: retry on -EAGAIN
Mailbox framework makes requests to the underlying drivers by holding a
spinlock. If the interrupt handler and the mailbox request happen to
occur on the same CPU, then the interrupt handler would never get a
chance to run, until the write to the controller is complete.

In some interrupt driven controllers, the interrupt handler is the one
to mark the transaction as complete and the controller status to ready
If the interrupt handler did not get a chance to run, then the mailbox
write would perpertually spin waiting on the controller to be free to
write the new request.

To avoid this kind of a deadlock, relax the CPU by relinquishing the
spinlock and retry again. This would give the interrupt handler a chance
to run and mark the controller state as idle before mailbox retries
sending the request again.

Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>