CHROMIUM: platform: arm: add event log support.
An in memory buffer is set up covering a range of addresses specified in the
device tree, and when a panic happens, an event log event is generated and
stored in that buffer. When the system reboots as a result of the panic, the
firmware discovers the event and records it in the event log in flash.

Signed-off-by: Gabe Black <gabeblack@google.com>
(Squashed related commits bc14076e8807 .. 6addb52fb680)
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>