drivers: mailbox: qcom-rpmh: add RPMH mailbox controller for QCOM SoC
Add mailbox controller driver for QCOM SoC's that have hardware based
shared resource management. The hardware IP known as RSC (Resource State
Coordinator) houses a block (DRV) is used to send requests from an
execution environment to shared resources housed in the common IP and
shared with other processors in the SoC.

The Trigger Control Set (TCS) is a mailbox slot that can house multiple
resource state requests and when triggered will issue these requests
through the proprietary bus to the Resouce Power Manager Hardened (RPMH)
blocks. These hardware blocks are capable of adjusting clocks, voltages
etc. The resource state request from Linux are aggregated along with
state requests from other processors in the SoC and the aggregate value
is applied on the resource. Every resource has a unique address and
supports pre-defined set of valid values for its states. The address and
the values are written to these TCS registers and triggered using a
register write indicating to the hardware block that a new resource
state is requested. The hardware responds by aggreating the state
requested along with state requests from other processors in the SoC and
applies the new value to the resource.

Resource requests made while the CPU is running are called active state
request and get a response from the hardware as an interrupt. The
interrupt response signals the completion of the requests at the RPMH
and the driver notifies the mailbox framework using the tx_done
callbacks. Requests sent when the processor is idle, to reduce resource
utilization are called sleep state requests and wake state requests
restore the resource states to previously requested values before
bringing the processor out of its low power state.

We support only active mode requests in this patch. The following
patches will add support for sleep and wake requests.

Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>