drivers: qcom: rpmh: cache sleep/wake state requests
Active state requests are sent immediately to the mailbox controller,
while sleep and wake state requests are cached in this driver to avoid
taxing the mailbox controller repeatedly. The cached values will be sent
to the controller when the rpmh_flush() is called.

Generally flushing is a system activity and may be called from the
system PM drivers when the system is entering suspend or deeper sleep
modes during cpuidle.

Also, allow invalidating the cached requests, so they be re-populatd
again.

Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>