drivers: qcom: rpmh: add support for batch RPMH request
Platform drivers who request a lot of resources at the same time, may
want to give RPMH library a batch of requests to be sent and wait on the
whole transaction to be complete. Generally, this happens when a usecase
is started, necessiating the drivers to turn a bunch of clocks and set
frequency requests on a few busses.

rpmh_write_batch() is a blocking call that can be used to send multiple
RPMH command sets. Each RPMH command set is set asynchronously and the
API blocks until all the command sets are complete and recieve the
tx_done callbaks.

signed-off-by: Lina Iyer <ilina@codeaurora.org>

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>