drivers: qcom: rpmh: disallow active votes in solver mode
The RSC hardware may be configured to act indepent of Linux, (firmware
control) while Linux is still running. This mode is called the 'solver'
mode for the RSC.

A separate RSC controller driver can make this decision and let RPMH
drivers know that the controller is not in the right state to support
sending active mode requests. Hence check for controller state before
issuing active state requests.

Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>