UPSTREAM: iommu/iova: Use raw_cpu_ptr() instead of get_cpu_ptr() for ->fq
get_cpu_ptr() disabled preemption and returns the ->fq object of the
current CPU. raw_cpu_ptr() does the same except that it not disable
preemption which means the scheduler can move it to another CPU after it
obtained the per-CPU object.
In this case this is not bad because the data structure itself is
protected with a spin_lock. This change shouldn't matter however on RT
it does because the sleeping lock can't be accessed with disabled
preemption.

Cc: Joerg Roedel <joro@8bytes.org>
Cc: iommu@lists.linux-foundation.org
Reported-by: vinadhy@gmail.com
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
(cherry picked from commit 94e2cc4dba39efbb63e84cd4f7243627c16ceac5)

BUG=b:74036927
TEST=Build kernel for cheza

Change-Id: I9fe86cc0d984f65413250196fb2cb6b6c45932ea
Signed-off-by: Vivek Gautam <vivek.gautam@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/1010047
Reviewed-by: Tomasz Figa <tfiga@chromium.org>