FROMLIST: clk: qcom: remove redundant PLL_MODE macro offset
The PLL_MODE offset macro is redundant which is defined as zero.
The offset in PLL structure is the address of PLL_MODE register
itself so the PLL_MODE can be removed. It will help in subsequent
patches to support different PLL offset registers to reduce the
code diff.

Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>
Patch-mainline: https://patchwork.kernel.org/patch/9976577/
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>