FROMLIST: clk: qcom: add and use alpha register width from PLL properties
Currently SUPPORTS_16BIT_ALPHA flag determines the PLL alpha
register width. If this flag is set then the alpha register width
is 16 bits otherwise it is 40 bits. The alpha width is always
fixed for PLL type so it can be added in PLL properties and clock
driver don’t have to specify explicitly.

The SUPPORTS_16BIT_ALPHA flag is unused in the current code so
it’s safe to remove this flags.

Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>
Patch-mainline: https://patchwork.kernel.org/patch/9976561/
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>