CHROMIUM: esdfs: Allow non-owners to touch
This modifies the permission checks in setattr to
allow for non-owners to modify the timestamp of
files to things other than the current time.
This still requires write access, as enforced by
the permission call, but relaxes the requirement
that the caller must be the owner, allowing those
with group permissions to change it as well.

BUG=b:63876697
TEST=compilation, mount esdfs

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 11118565
Bug: 34948874
Change-Id: Ic10b6a0819727ed26a8cd53ee8e94a4e59acbd82
Reviewed-on: https://chromium-review.googlesource.com/791983
Commit-Ready: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
(cherry picked from commit cb965b4c1f3987d9287263b6c1d96cd056e70709)
Reviewed-on: https://chromium-review.googlesource.com/963418