FROMLIST: clk: qcom: clear hardware clock control bit of RCG
For upcoming targets, the hardware clock control bit is set for most of
root clocks which needs to be cleared for software to be able to control
those root clocks. For older targets like MSM8996, this bit is reserved
bit and having POR value as 0 so this patch will work for the older
targets too. So update the configuration mask to take care of the same to
clear hardware clock control bit.

Change-Id: I367e0bd2c215391c0058731cc45dea259711a643
Signed-off-by: Amit Nischal <anischal@codeaurora.org>
Patch-mainline: https://patchwork.kernel.org/patch/10033651/
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>