FROMLIST: gpiolib: add bitmask for valid GPIO lines
Add support for specifying that some GPIOs within a range are unavailable.
Some systems have a sparse list of GPIOs, where a range of GPIOs is
specified (usually 0 to n-1), but some subset within that range is
absent or unavailable for whatever reason.

To support this, allow drivers to specify a bitmask of GPIOs that
are present or absent.  Gpiolib will then block access to those that
are absent.

Signed-off-by: Timur Tabi <timur@codeaurora.org>
Patch-mainline: https://patchwork.kernel.org/patch/10033139/
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>