UPSTREAM: soc: qcom: Introduce QMI helpers
Drivers that needs to communicate with a remote QMI service all has to
perform the operations of discovering the service, encoding and decoding
the messages and operate the socket. This introduces an abstraction for
these common operations, reducing most of the duplication in such cases.

Acked-by: Chris Lew <clew@codeaurora.org>
Tested-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>

BUG=None
TEST=Build and boot

Change-Id: I383ed8b3b8dbf04944f35f361f7fb1a37459b493
Signed-off-by: Sibi S <sibis@codeaurora.org>
(cherry picked from commit 3830d0771ef66a3bb9ab57b70630b3598593a4f0)