INTERNAL: ARM: dts: msm: Introduce topology for sdm845
Topology to describe Network On Chip(NOC) connectivity for sdm845.
This is used to map the paths between masters and slaves across the
internally connected buses. Each node represents a point of aggregation
and contains details such as width, and priority configuration to help
provide sufficient bandwidth to each client.

Change-Id: I6ab94b4da3036b1854e4991e93abe2dadf03d132
Signed-off-by: David Dai <daidavid1@codeaurora.org>