UPSTREAM: drm/msm: Trigger fence completion from GPU
Interrupt commands causes the CP to trigger an interrupt as the command
is processed, regardless of the GPU being done processing previous
commands. This is seen by the interrupt being delivered before the
fence is written on 8974 and is likely the cause of the additional
CP_WAIT_FOR_IDLE workaround found for a306, which would cause the CP to
wait for the GPU to go idle before triggering the interrupt.

Instead we can set the (undocumented) BIT(31) of the CACHE_FLUSH_TS
which will cause a special CACHE_FLUSH_TS interrupt to be triggered from
the GPU as the write event is processed.

Add CACHE_FLUSH_TS to the IRQ masks of A3xx and A4xx and remove the
workaround for A306.

Suggested-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Rob Clark <robdclark@gmail.com>
(cherry picked from commit 79d57bf6fa3bcc0ec5fc3b8140c4df1d696f593b)
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Git-commit: f2ab10dabf6ebc373b779913394a016552e58085
Git-repo: https://gitlab.freedesktop.org/seanpaul/dpu-staging
Signed-off-by: Rajesh Yadav <ryadav@codeaurora.org>

Change-Id: I7f5dc941af8a145b611caa8eb9ffa6a534449580