FROMLIST: soc: qcom: Add APR bus driver
This patch adds support toi APR bus (Asynchronous Packet Router) driver.
APR driver is made as a bus driver so that the apr devices can added
and removed more dynamically depending on the state of the services on
the dsp. APR is used for communication between application processor and
QDSP to use services on QDSP like Audio and others.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>

BUG=chromium:795946
TEST=Build and boot
(am from https://patchwork.kernel.org/patch/10272905/)

Change-Id: I35f7148a587e234d5f7fbc679ad567f09409c1b7
Signed-off-by: Rohit kumar <rohitkr@codeaurora.org>