CHROMIUM: tpm: Restore quality of tpm-rng to 1000
This change reapplies the applicable portion of
commit d0621b4cd498 ("CHROMIUM: tpm-rng: Make hwrng working")

The idea is to set a non-zero quality in the tpm-rng, so that
the hwrng_fill background thread gets started. This thread is able
to fill the randomness pool in the background, which helps speed up
boot time.  The original change contains a bit more context.

BUG=chromium:673183, b:78841590
TEST=Build kernel for Cheza, though no TPM yet.
TEST=Deploy grunt kernel, UI comes up to login screen in < 15 seconds.

Change-Id: I2bd31a6fe89bf979353c0c668dbdacd92d3731a6
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/967183
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>