INTERNAL: tty: qcom_geni_serial: Fix power state and add high-speed rates
The PM state for all ports is initialized to undefined and moves to on
when the port is first added. The driver currently only honors power state
changes going from on to off and vice-versa.  Modify this to include
undefined also, but make this conditional to non-console ports as the
framework doesn't move the port's state back to off for console ports
after the initial setup.

Also add support for higher baud rates.

Change-Id: Ib1ee5cba9bcb8ec1943212277a394e0ab4e666f4
Signed-off-by: Girish Mahadevan <girishm@codeaurora.org>