CHROMIUM: platform: chrome: drop chromeos_arm driver
This driver is no longer needed. It only served to:
 (a) register sysfs shortcuts for a few GPIOs
 (b) register a panic handler for stashing eventlog info in memory (?)

Item (a) is supported via named GPIO lookups (/dev/gpiochip* chardev
API), and we'd like to stop using this driver for that.

Item (b) was only used on Peach AFAICT. We could probably resurrect the
driver and do this differently, if that's really the way this should go.

BUG=chromium:897992
TEST=boot, `crossystem wpsw_cur` works without this driver

Change-Id: If6c0bc0140134dece306833a3717e84056a77dea
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1297837
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>