CHROMIUM: esdfs: do not modify permissions on Downloads directory
On lookup, esdfs touches up the ownership of the lower filesystem.
Ignore the Downloads directory while setting permissions, we assume
that the Downloads directory would have the correct permissions
set up when the user logs in. Note that this only skips the
Downloads directory: all subdirectories and files would still be
get their lower ownership changed on esdfs_lookup.

Also, ATTR_KILL_S{U,G}ID is called for inodes that are not
directories in esdfs_derive_lower_ownership, which can also be
skipped for the special Downloads directory.

BUG=chromium:916015
TEST=Downloads gid does not change to "chronos".

Change-Id: I09199fa381f450aa9a4c1b52d5e93ef46466462e
Signed-off-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1385394
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>