FROMLIST: sched: Add core wide task selection and scheduling.
Instead of only selecting a local task, select a task for all SMT
siblings for every reschedule on the core (irrespective which logical
CPU does the reschedule).

There could be races in core scheduler where a CPU is trying to pick
a task for its sibling in core scheduler, when that CPU has just been
offlined.  We should not schedule any tasks on the CPU in this case.
Return an idle task in pick_next_task for this situation.

NOTE: there is still potential for siblings rivalry.
NOTE: this is far too complicated; but thus far I've failed to
      simplify it further.

BUG=b:152605392
TEST=run power_VideoCall test

(am from https://lore.kernel.org/patchwork/patch/1204980/)

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Julien Desfossez <jdesfossez@digitalocean.com>
Signed-off-by: Vineeth Remanan Pillai <vpillai@digitalocean.com>
Signed-off-by: Aaron Lu <aaron.lu@linux.alibaba.com>
Signed-off-by: Tim Chen <tim.c.chen@linux.intel.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: I5b85c96edc3218f48dd583448adc76be35266033
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2318718
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>