BACKPORT: FROMLIST: ath10k: Fix NSS tx stats for legacy rates
The ratecode field given by firmware in the tx stats
gives the details of NSS and MCS, if the preamble
specifies that it is HT/VHT. In case of OFDM/CCK, as
specified by the preamble, the ratecode indicates the
legacy rate instead of NSS/MCS data. For OFDM/CCK cases,
the NSS should always be 1, i.e. 1x1 mode for 11g and 11b.

Currently the legacy rate is misinterpreted as NSS & MCS, leading
to incorrect updation of tx stats corresponding to NSS, which is
displayed in the tx stats via debugfs.

Fix the nss value to 1 for 11b and 11g modes by checking
the preamble indication in the ratecode provided by the firmware.

Tested HW: WCN3990
Tested FW: WLAN.HL.3.1-00784-QCAHLSWMTPLZ-1

am link: https://patchwork.kernel.org/patch/10847737/

Fixes: e88975ca37d1 ("ath10k: dump tx stats in rate table format")
Change-Id: Ieb3e244620d6e4065bf3a17f1a78b62bbfaafd04
Cc: ath10k.lists.infradead.org, linux-wireless@vger.kernel.org
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>