BACKPORT: mac80211: probe unexercised mesh links
The requirement for mesh link metric refreshing, is that from one
mesh point we be able to send some data frames to other mesh points
which are not currently selected as a primary traffic path, but which
are only 1 hop away. The absence of the primary path to the chosen node
makes it necessary to apply some form of marking on a chosen packet
stream so that the packets can be properly steered to the selected node
for testing, and not by the regular mesh path lookup.

Tested-by: Pradeep Kumar Chitrapu <pradeepc@codeaurora.org>
Signed-off-by: Rajkumar Manoharan <rmanohar@codeaurora.org>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(cherry picked from commit 8828f81ad4a2f4e89ebe6e7793c06ed767c31d53)

BUG=b:134764458
TEST=manually test and verified by sniffer on mistral
Signed-off-by: Roy Luo <royluo@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1650690
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Julan Hsu <julanhsu@google.com>
Git-repo: https://chromium.googlesource.com/a/chromiumos/third_party/kernel
Git-commit: 51d60bc9eaae08e2f8ace310ed27fdfb743cd1cc

Change-Id: I4cb71f544aa2ce98f4bf096979e1b05380a588e8
Signed-off-by: Bikshapathi Kothapeta <bkotha@codeaurora.org>