CHROMIUM: iio: cros_ec: Clamp timestamps in sensor ring
With sensors enabled (usually at higher sample rates) we find
about 1 out of 500 samples that will have an event timestamp in
the future. This change clamps the timestamps to never exceed
the time they arrive at the sensor ring.

With the logging enabled, during the CTS pass, we got 5 prints
(500-599 instances) with an average of about 120us shaved.

BUG=b:129159505,b:123700100,b:130281042
BRANCH=None
TEST=Ran Android CTS

Change-Id: Icc2d27940f3bede4006bd606d620d9335d38f86f
Signed-off-by: Yuval Peress <peress@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1600418
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>