CHROMIUM: WAR: ath10k: Set peer flags correctly in peer assoc cmd
WMI_TLV_PEER_AUTH flag in peer assoc cmd indicates to firmware
that peer is authorised to send data. Currently, the WMI_TLV_PEER_AUTH
flag is set unconditionally, this leads to improper behavior of
firmware when vdev of type AP is brought up in secure mode as in secure
mode data frames should be sent only after key exchange and installation
is done. Currently, as in peer assoc cmd we set WMI_TLV_PEER_AUTH flag,
it indicates to firmware that connection is authorised and it allows the
data frames to be sent before the keys are not installed in open mode.

To mitigate this issue don't set the the WMI_TLV_PEER_AUTH in peer
assoc cmd if ptk or gtk key exchange is to be done after association.
This will indicate firmware that connection to the peer is not authorised
and it will not send data frames to the peer till the keys are installed.

Change-Id: I585cfe0fa855883cf39e01866e198fad37a5088f
Signed-off-by: Surabhi Vishnoi <svishnoi@codeaurora.org>