FROMLIST: tty: serial: msm_serial: Fix XON/XOFF
When the tty layer requests the uart to throttle, the current code
executing in msm_serial will trigger "Bad mode in Error Handler" and
generate an invalid stack frame in pstore before rebooting (that is if
pstore is indeed configured: otherwise the user shall just notice a
reboot with no further information dumped to the console).

This patch replaces the PIO byte accessor with the word accessor
already used in PIO mode.

Fixes: 68252424a7c7 ("tty: serial: msm: Support big-endian CPUs")
Cc: stable@vger.kernel.org

(am from https://patchwork.kernel.org/patch/10952209/)

BUG=b:none
TEST=Flood console or input more than 8Kb data

Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Bikshapathi Kothapeta <bkotha@codeaurora.org>
Change-Id: If50b8833eddc0c4ce4b4c456e2a1ec1a42d32281
Signed-off-by: Praveen Koya <pkoya@codeaurora.org>