BACKPORT:UPSTREAM: Bluetooth: hci_qca: wcn3990: Drop baudrate change vendor event
Firmware download to the WCN3990 often fails with a 'TLV response size
mismatch' error:

[  133.064659] Bluetooth: hci0: setting up wcn3990
[  133.489150] Bluetooth: hci0: QCA controller version 0x02140201
[  133.495245] Bluetooth: hci0: QCA Downloading qca/crbtfw21.tlv
[  133.507214] Bluetooth: hci0: QCA TLV response size mismatch
[  133.513265] Bluetooth: hci0: QCA Failed to download patch (-84)

This is caused by a vendor event that corresponds to an earlier command
to change the baudrate. The event is not processed in the context of the
baudrate change and is later interpreted as response to the firmware
download command (which is also a vendor command), but the driver detects
that the event doesn't have the expected amount of associated data.

More details:

For the WCN3990 the vendor command for a baudrate change isn't sent as
synchronous HCI command, because the controller sends the corresponding
vendor event with the new baudrate. The event is received and decoded
after the baudrate change of the host port.

Identify the 'unused' event when it is received and don't add it to
the queue of RX frames.

(cherry picked from commit 2faa3f15fa2fc13ad1e12a61fc3dbb14b326ba7a)

Conflicts:
fixed
  --- a/drivers/bluetooth/hci_qca.c
  +++ b/drivers/bluetooth/hci_qca.c
  - +static int qca_recv_event(struct hci_dev *hdev, struct sk_buff *skb)
  - +{
  - +	struct hci_uart *hu = hci_get_drvdata(hdev);
  + +        struct hci_uart *hu = hci_get_drvdata(hdev);
  - +	struct qca_data *qca = hu->priv;
  + +        struct qca_data *qca = hu->priv;
  - +	return hci_recv_frame(hdev, skb);
  - +}
  - +
  - -	{ H4_RECV_EVENT,           .recv = hci_recv_frame    },
  - +	{ H4_RECV_EVENT,           .recv = qca_recv_event    },

BUG=b:150161602
TEST=bring hci0 up/down repeatedly

(cherry picked from commit ee36642876b5303818368f5064c524faec02b808
https://chromium.googlesource.com/chromiumos/third_party/kernel/+/refs/heads/chromeos-4.14-gw)
Change-Id: Ifb36891f687c3d1a72a8a9891b88e4f9f97a9983
Signed-off-by: Harish Bandi <c-hbandi@codeaurora.org>
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2100350
Reviewed-by: SANTHOSH JANARDHANA HASSAN <sahassan@google.com>
Tested-by: SANTHOSH JANARDHANA HASSAN <sahassan@google.com>
Commit-Queue: Julan Hsu <julanhsu@google.com>
Signed-off-by: vamsi singamsetty <vamssi@codeaurora.org>

Change-Id: I3d272bc3b5200d07844e01c67f31f07636d932c7