CHROMIUM: platform/x86: intel_pmc_core: dev_warn() on S0ix failure
This is to maintain the status quo with what we already have in tree,
before migrating to the upstream version of patches.

While submitting the patches, the maintainer had given a number of
suggestions including to convert this into a debug messages. I
decided not to be adamant about this, in support of letting other
patches be accepted. However we do need this to be pr_warn in order
to see the logs in case of S0ix failures. This gets printed in really
rare conditions, but we really would like a log for those rare
situations.

Signed-off-by: Rajat Jain <rajatja@google.com>
BUG=None
TEST=Verify S0ix failure
Change-Id: I3c02713de9655635c370bded0b7479990d5f2a69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1681415
Reviewed-by: Furquan Shaikh <furquan@chromium.org>