BACKPORT: FROMLIST: clk: Remove cached cores in parent map during unregister
As clocks are registered their parents are resolved and the parent_map
is updated to cache the clk_core objects of each existing parent.
But in the event of a clock being unregistered this cache will carry
dangling pointers if not invalidated, so do this for all children of the
clock being unregistered.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Tested-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>
(am from https://lore.kernel.org/patchwork/patch/1103323/)
(also found at https://lkml.kernel.org/r/20190723051446.20013-1-bjorn.andersson@linaro.org)

BUG=None
TEST=Boot cheza and no kernel oops

Change-Id: Id0c742eede0c739a46506aa431130d0e685d147a
[groeck: backport: New caller of clk_core_set_parent_nolock() in upstream kernel]
Signed-off-by: Guenter Roeck <groeck@chromium.org>